requirements.txt
flask
flask-cors
g4f
gunicorn
requests
python-dotenv
